update merge event workflow to ...
make the autotweet step optional, in the event the twitter post fails for some reason, such as when a user edits a blog that was recently posted and a redundant tweet is sent, twitter may deny the redundant tweet as duplicate. Adding continue-on-error: true to the autotweet step should ensure the entire flow doesnt fail if problem with the autotweet step.